Add common provider
