+++ api/v1beta1/spiaccesstoken_types.go
@@ -82,7 +82,7 @@ type ServiceProviderType string
 const (
 	ServiceProviderTypeGitHub ServiceProviderType = "GitHub"
 	ServiceProviderTypeQuay   ServiceProviderType = "Quay"
-	ServiceProviderTypeSnyk   ServiceProviderType = "Snyk"
+	ServiceProviderTypeCommon ServiceProviderType = "Common"
 )
 
 // Permission is an element of Permissions and express a requirement on the service provider scopes in an agnostic
