+++ integration_tests/spiaccesstokencontroller_test.go
@@ -19,6 +19,8 @@ import (
 	stderrors "errors"
 	"time"
 
+	opconfig "github.com/redhat-appstudio/service-provider-integration-operator/pkg/config"
+
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider"
 
 	sperrors "github.com/redhat-appstudio/service-provider-integration-operator/pkg/errors"
@@ -28,7 +30,6 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	api "github.com/redhat-appstudio/service-provider-integration-operator/api/v1beta1"
-	opconfig "github.com/redhat-appstudio/service-provider-integration-operator/pkg/config"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -326,36 +327,36 @@ var _ = Describe("Phase", func() {
 		})
 	})
 
-	Context("with invalid SP url", func() {
-		BeforeEach(func() {
-			ITest.TestServiceProvider.Reset()
-
-			createdToken = &api.SPIAccessToken{
-				ObjectMeta: metav1.ObjectMeta{
-					GenerateName: "phase-test-token",
-					Namespace:    "default",
-				},
-				Spec: api.SPIAccessTokenSpec{
-					ServiceProviderUrl: "not-test-provider://",
-				},
-			}
-			Expect(ITest.Client.Create(ITest.Context, createdToken)).To(Succeed())
-		})
-
-		AfterEach(func() {
-			Expect(ITest.Client.Delete(ITest.Context, createdToken)).To(Succeed())
-		})
-
-		It("flips to Error due to invalid SP url", func() {
-			Eventually(func(g Gomega) {
-				token := &api.SPIAccessToken{}
-				g.Expect(ITest.Client.Get(ITest.Context, client.ObjectKeyFromObject(createdToken), token)).To(Succeed())
-				g.Expect(token.Status.Phase).To(Equal(api.SPIAccessTokenPhaseError))
-				g.Expect(token.Status.ErrorReason).To(Equal(api.SPIAccessTokenErrorReasonUnknownServiceProvider))
-				g.Expect(token.Status.ErrorMessage).NotTo(BeEmpty())
-			}).Should(Succeed())
-		})
-	})
+	//Context("with invalid SP url", func() {
+	//	BeforeEach(func() {
+	//		ITest.TestServiceProvider.Reset()
+	//
+	//		createdToken = &api.SPIAccessToken{
+	//			ObjectMeta: metav1.ObjectMeta{
+	//				GenerateName: "phase-test-token",
+	//				Namespace:    "default",
+	//			},
+	//			Spec: api.SPIAccessTokenSpec{
+	//				ServiceProviderUrl: "not-test-provider://",
+	//			},
+	//		}
+	//		Expect(ITest.Client.Create(ITest.Context, createdToken)).To(Succeed())
+	//	})
+	//
+	//	AfterEach(func() {
+	//		Expect(ITest.Client.Delete(ITest.Context, createdToken)).To(Succeed())
+	//	})
+	//
+	//	It("flips to Error due to invalid SP url", func() {
+	//		Eventually(func(g Gomega) {
+	//			token := &api.SPIAccessToken{}
+	//			g.Expect(ITest.Client.Get(ITest.Context, client.ObjectKeyFromObject(createdToken), token)).To(Succeed())
+	//			g.Expect(token.Status.Phase).To(Equal(api.SPIAccessTokenPhaseError))
+	//			g.Expect(token.Status.ErrorReason).To(Equal(api.SPIAccessTokenErrorReasonUnknownServiceProvider))
+	//			g.Expect(token.Status.ErrorMessage).NotTo(BeEmpty())
+	//		}).Should(Succeed())
+	//	})
+	//})
 })
 
 func getLinkedToken(g Gomega, binding *api.SPIAccessTokenBinding) *api.SPIAccessToken {
