+++ pkg/serviceproviders/known.go
@@ -16,9 +16,9 @@ package serviceproviders
 
 import (
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider"
+	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider/common"
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider/github"
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider/quay"
-	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider/snyk"
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/spi-shared/config"
 )
 
@@ -31,6 +31,6 @@ func KnownInitializers() map[config.ServiceProviderType]serviceprovider.Initiali
 	return map[config.ServiceProviderType]serviceprovider.Initializer{
 		config.ServiceProviderTypeGitHub: github.Initializer,
 		config.ServiceProviderTypeQuay:   quay.Initializer,
-		config.ServiceProviderTypeSnyk:   snyk.Initializer,
+		config.ServiceProviderTypeCommon: common.Initializer,
 	}
 }
